% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-marimekko.R
\docType{data}
\name{marimekko_pal}
\alias{marimekko_pal}
\title{Unikko-inspired colour palette}
\format{
An object of class \code{character} of length 8.
}
\usage{
marimekko_pal
}
\description{
A character vector of 8 bold colours inspired by Marimekko's iconic
Unikko poppy pattern. Vibrant, high-contrast tones suited for
categorical data visualisation.
}
\keyword{datasets}
