% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-marimekko.R
\name{scale_colour_marimekko}
\alias{scale_colour_marimekko}
\title{Marimekko colour scale}
\usage{
scale_colour_marimekko(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2:scale_manual]{ggplot2::scale_colour_manual()}}.}
}
\value{
A ggplot2 scale.
}
\description{
A discrete colour scale using earthy Nordic-inspired colours.
}
\examples{
library(ggplot2)

titanic <- as.data.frame(Titanic)
ggplot(titanic) +
  geom_marimekko(
    aes(fill = Survived, weight = Freq),
    formula = ~ Class | Survived
  ) +
  scale_colour_marimekko()

}
