% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-x-marimekko.R
\name{scale_x_marimekko}
\alias{scale_x_marimekko}
\title{X-axis scale for marimekko plots}
\usage{
scale_x_marimekko(show_percentages = FALSE, ...)
}
\arguments{
\item{show_percentages}{Logical. If \code{TRUE}, appends marginal
percentage to each x-axis label. Default \code{FALSE}.}

\item{...}{Arguments passed to \code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}}.}
}
\value{
A ggplot2 scale.
}
\description{
X-axis scale for marimekko plots
}
\examples{
library(ggplot2)

titanic <- as.data.frame(Titanic)
ggplot(titanic) +
  geom_marimekko(
    aes(fill = Survived, weight = Freq),
    formula = ~ Class | Survived
  ) +
  scale_x_marimekko(show_percentages = TRUE)

}
