% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-y-marimekko.R
\name{scale_y_marimekko}
\alias{scale_y_marimekko}
\title{Y-axis scale for marimekko plots}
\usage{
scale_y_marimekko(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}}.}
}
\value{
A ggplot2 scale.
}
\description{
Y-axis scale for marimekko plots
}
\examples{
library(ggplot2)

titanic <- as.data.frame(Titanic)
ggplot(titanic) +
  geom_marimekko(
    aes(fill = Survived, weight = Freq),
    formula = ~ Class | Survived
  ) +
  scale_x_marimekko() +
  scale_y_marimekko()

}
