% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mekko-package.R
\docType{package}
\name{marimekko-package}
\alias{marimekko}
\alias{marimekko-package}
\title{marimekko: marimekko and Mosaic Plots for 'ggplot2'}
\description{
Create marimekko (mosaic) plots as a 'ggplot2' layer. Column widths encode marginal proportions of one categorical variable and segment heights encode conditional proportions of a second categorical variable. Provides a simpler alternative to 'ggmosaic'.
}
\author{
\strong{Maintainer}: Dawid Kaledkowski \email{dawid.kaledkowski@gmail.com}

}
\keyword{internal}
