% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-marimekko.R
\docType{data}
\name{marimekko_pal}
\alias{marimekko_pal}
\title{Earthy Nordic colour palette}
\format{
An object of class \code{character} of length 8.
}
\usage{
marimekko_pal
}
\description{
A character vector of 8 earthy, Nordic-inspired colours for use with
marimekko plots. Muted natural tones that work well for categorical
data visualisation.
}
\keyword{datasets}
